refactored CacheBackedEmbeddings
