+++ libs/community/tests/unit_tests/embeddings/test_imports.py
@@ -4,6 +4,7 @@ EXPECTED_ALL = [
     "OpenAIEmbeddings",
     "AzureOpenAIEmbeddings",
     "BaichuanTextEmbeddings",
+    "CacheBackedEmbeddings",
     "ClarifaiEmbeddings",
     "CohereEmbeddings",
     "DatabricksEmbeddings",
